Project2/cse546ir/requirements.txt
boto3==1.21.44
botocore==1.24.44
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.32.0
idna==3.3
jmespath==1.0.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
